agents/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM Providers
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Databricks Integration
databricks-sdk>=0.18.0
mlflow>=2.10.0
pyspark>=3.5.0

# Vector Stores and Memory
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Tools and Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0

# Optional Voice/Vision capabilities
# elevenlabs>=0.2.0
# openai>=1.0.0  # For vision/whisper

# Development
pytest>=7.4.0
black>=23.0.0

